Fix texture-only arg in combine samplers transform.
When a function has an unused texture-only or sampler-only argument,
and no used texture/sampler pairs, the CombineSamplers transform was
failing to remove it from the argument list. The reason is an
incorrect early-out when no texture/sampler pairs are found.

https://dawn-review.googlesource.com/c/dawn/+/154181 attempted to
fix this by adding the unused arguments back as texture/sampler
solo pairs in CollectTextureSamplerPairs(). However, this is
problematic for a number of reasons.

This CL essentially reverts that change, and removes the incorrect
early-out, causing unused texture and sampler parameters to be
(correctly) removed.

Bug: chromium:356298370
Change-Id: If69e4d69dbf308637689622094193a285f9c3a46
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/202155
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Stephen White <senorblanco@chromium.org>