[tint][ir][val] Check disards in validator
Checks the number of results and operands, as well that discard is not
called outside of fragment entry points.

The top-level Run() for the validator has been refactored to make it
clearer what is run as a first-pass validation that ensures the
structural soundness of the IR, and what are later checks that depend
on this first pass completing before running.

Fixes: 359593395

Change-Id: I042ec47b9754aa3b6da44d667aa9c0147af3da9d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/202414
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>