[benchmark] Convert SPIR-V when generating header
Update the benchmark header generation script to invoke Tint to
convert SPIR-V shaders to WGSL, and have it always generate the header
at the checked-in location.

Remove the GN target (because outputs cannot be in the source
directory), but leave the CMake target for convenience.

Bug: 42251293
Change-Id: I29baafa4890604540050fa02a023f6796f5b0b9b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/202359
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>