[tint] Implement quadBroadcast
Implement WGSL builtin quadBroadcast in the SPIR-V IR backend, MSL
AST and IR backends and HLSL AST and IR backends. Subgroup builtins
are not yet implemented for the GLSL backend.

Constness of the ID arg is validated in the resolver, and a polyfill is
implemented in the SPIR-V backend to bitcast signed integer IDs to
unsigned.

Validation against negative ID args has been added for both
subgroupBroadcast and quadBroadcast.

Bug: 354738715
Change-Id: Iadb9726113bb964d51bbf4788f5deae4ae85b6d5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/202374
Reviewed-by: dan sinclair <dsinclair@google.com>
Commit-Queue: Natalie Chouinard <chouinard@google.com>