[tint] validate clamp const low <= const high when first arg is runtime-eval
Generate an error for clamp(x,low,high) when low and high are
constant and low <= high. The new functionality is that we
check this even when x is a runtime-expression.

Added a unit test for the const case.
Added end2end tests for const and override cases, since the override
case needs substitute-overrides to run.

Bug: chromium:351378281
Change-Id: Iaeca78e08ded5b7b17469a56a617db9e5f730c18
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/202215
Commit-Queue: David Neto <dneto@google.com>
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: David Neto <dneto@google.com>