[benchmark] Add --check-stale option
Adds a flag to the benchmark header generation script which can be
used to check if the generated header is stale or not. This can then
be used by CQ to automatically detect when the benchmark header needs
to be regenerated.

Bug: 42251293
Change-Id: I43b590512dc987e98afbf0759cac34799286f6ec
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/202362
Auto-Submit: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>