EncodingContext: Use a Status enum to handle the various states.
Following the logic of what's happening in EncodingContext was difficult
as there was a mix of booleans, mTopLevelEncoder != nullptr and implicit
handling of when EncodingContext was used for an error encoder.

Introduce a status enum with the invariant that
(mStatus == Open) == (mToplevelEncoder != nullptr) to keep the fast
check for ValidateCanEncodeOn (renamed from CheckCurrentEncoder).

Adds a constructor for EncodingContext with the ErrorTag that's used in
the error constructors of CommandEncoder and RenderBundleEncoder.

Removes dawn_unittests expliclty calling EncodingContext::Destroy at
random times which forced more complex handling than necessary
(Destroy is called only in the top level encoder's DestroyImpl or in
~EncodingContext).

Slightly simplify the handling of the acquisition of the CommandIterator
as it can only be done once, and can always go through AcquireCommands.

Fixup a couple tests for the functional change that now starting a pass
on an error CommandEncoder will emit a device error instead of silently
failing.

Bug: 355308275
Change-Id: I5536f17da10b157c2924372ff499387743e0a485
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/200135
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>