Add D3D12 Backend for MultiDrawIndirect Feature
This CL adds support for the D3D12 backend, which requires emulation of
base vertex and base instance parameters in the indirect arguments.
The indirect draw validation compute pass duplicates the parameters if it is used in the shaders.
MultiDraw capability is supported on all devices with D3D12.
New tests added for MultiDrawIndexedIndirect to test baseVertex and firstInstance.

Change-Id: I75bc48243e4801f49e6e50091cf1560593a3d14c
Bug: 356461286
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/203254
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Srijan Dhungana <srijan.dhungana6@gmail.com>
Reviewed-by: Austin Eng <enga@chromium.org>