[benchmark] Generate header at build time
Instead of generating the entire header offline and checking it in,
just generate the WGSL shaders for any SPIR-V inputs and check those
in instead. Move the header generation back to build time so that we
do not have to check it in.

This makes future diffs for updated WGSL codegen actually readable.

Bug: 42251293
Change-Id: Ief4610404bec5d4ef9e79b230d51aaacad588f2d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/204234
Auto-Submit: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>