Create a shared utils folder.
This CL creates a `src/utils` folder and adds a `compiler.h` header. A
define is pulled out of the Dawn common compilers file and the tint
macros compiler file and added to the header.

The define is the same, just for `TINT_` or `DAWN_`. It will be
collapsed to a single define in a future CL.

Change-Id: Idb420cee2803a23353d3fa8b893db35fd7069206
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/203694
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>