[ir] Emit unused side-effecting expressions
When function calls are used by instructions that are themselves
unused, we were dropping the whole expression chain and never emitting
the function call at all.

Fix this in `ValueToLet` by making sure that all unused expressions
with side effects are associated with an instruction that will become
a statement (either a `call` or a `let`).

Fixed: 342650077
Fixed: 347690530
Fixed: 356419165
Change-Id: Ib78e31a4c32ddc02aa818316b00078ff80d961f8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/205116
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>