metal: Disable pixel local storage feature
The new MSL backend does not yet support the pixel local storage
extension, which is not currently used by Graphite when targeting the
Metal backend. We can re-enable this feature in the future when we
want to start using PLS on Metal for real.

Bug: 42251016
Change-Id: Ifd1a48b2c1f2c0c9e00670106a8ae0a4b3501e1a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/205420
Auto-Submit: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>