Support validating buffer usage with usage validation mode
This patch adds `UsageValidationMode` as the third parameter of
`ValidateCanUseAs` with a buffer as input to align with the
format of the function `ValidateCanUseAs` with a texture as input.

This patch also updates the test `InternalStorageBufferBindingTests`
by using classes in `native::` instead of the `wgpu::` to better
test the `internal usage` of the buffers.

Bug: chromium:42240463
Test: dawn_unittests
Change-Id: I2db2c922555ded050949da20828be17c95499152
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/204914
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>