SwapChainD3D12: Remember resource state for when reusing buffers.
Most applications would reconfigure the surface when there is no
current texture. However reconfiguration can also be done while a
texture is current and such that the IDXGISwapChain could recycle
buffers. In that case the ID3D12Resource of the current texture might
not be in D3D12_RESOURCE_STATE_COMMON like we assumed for all newly
returned wgpu::SurfaceTextures but a different one (based on what
barriers where done previously).

Add a state to d3d12::Texture::CreateForSwapChain as well as a
GetResourceStateForSwapChain getter and use it in d3d12::SwapChain to
keep track of the current state of the ID3D12Resource for a swap chain
buffer.

Bug: 42241264
Change-Id: I28745c8867ad12e3fa3d780e536e14247fd57db5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/205455
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Geoff Lang <geofflang@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>