Add Metal backend for Multi Draw Indirect
The Metal backend has to emulate the multi draws commands. It converts
and encodes the multi draws into an indirect command buffer.
This is done in a compute pass where it does the conversion and also
validates the arguments if needed. This would then skip the common
validation codepath that is implemented for D3D12 and Vulkan

Bug: 356461286.
Change-Id: Ieee5345d366441f250d56ad04d6c74aa04f94ccb
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/203934
Commit-Queue: Srijan Dhungana <srijan.dhungana6@gmail.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>