dawn::native: Fix the check for no attachment pipelines.
The check for attachments being required for render pipelines was
incorrect because it was only performed if desc->fragment != nullptr,
which is incorrect because desc->fragment == nullptr means that there is
no color attachment, and is an error if there's no depthStencil state.

Adds a regression test.
Fixes unittests that were triggering this validation by mistake, and
sent a CTS PR for the same purpose.

Bug: 341812551
Change-Id: I1daf726e258712535a42451323c84349257aefe4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/204934
Reviewed-by: Jiawei Shao <jiawei.shao@intel.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>