Add "hlsl-fxc" format to Tint executable
Treat "hlsl-fxc" as a separate format to "hlsl", producing HLSL for FXC,
while "hlsl" means produce HLSL for DXC. This is now used to set the
HLSL `Compiler` option accordingly.

Also, no longer validate HLSL with both DXC and FXC. If the format is
"hlsl-fxc", validate with FXC; if "hlsl", then validate with DXC. This
is simpler, and mirrors how it works for the other backends.

Regenerated Tint e2e tests with this change updates a few "ir.fxc.hlsl"
files now that they are generated using the FXC-enabled path.
Specifically, these run the FxcPolyfill transform, which implements a
workaround for default-only switch statements.

Bug: 369233346
Change-Id: Ic5ea8ee5e3c4a4c24accd5a4b3c6f82d29646d0f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/208015
Commit-Queue: Antonio Maiorano <amaiorano@google.com>
Reviewed-by: James Price <jrprice@google.com>