Use MacOS system clang instead of brew's
I don't know why current CI scripts try to use brew's LLVM instead of Apple's, but it turns out that simply switching to Apple's seems to fix the compilation issue that we've been having for months!

PS: Is there a plan of having some of these CI upload the artefacts to a proper GitHub Release, so that they do not expire? It is useful when using Dawn to be able to freeze the version that we rely on, but it is not reasonable to freeze to a binary artifact that won't last.

This is an imported pull request from
https://github.com/google/dawn/pull/21

GITHUB_PR_HEAD_SHA=d4f5d515d907850130b2f39b619af3cd9307ca79
ORIGINAL_AUTHOR=Elie Michel <eliemichel@users.noreply.github.com>
GitOrigin-RevId: bf40e7aa5459dde6699b04d69e8014ecf77bb20f
Change-Id: I62d70daf1b92307ee6e2f12a91720ecd52b54c5f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/207137
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: dan sinclair <dsinclair@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>