Fold wgpu::NullableStringView into wgpu::StringView
We decided that required StringViews would result in the empty string if
passed the nil StringView (NULL, WGPU_STRLEN). This means that the
NullableStringView concept is no longer needed and can be removed.

This is slightly involved as we need to remove the automatic conversion
from StringView to std::string, otherwise the automatic conversion to
std::optional<std::string> is ambiguous (either it is a direct
conversion, or first a conversion to std::string then to std::optional<...>).

Bug: 42241188
Change-Id: I24970aa8e71c34b9e7624f354f5bd92c42267f9d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/207994
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>