[Vulkan] Hardcode max possible number of Vk slots for YCbCr samplers
YCbCr samplers can take up multiple Vk descriptor slots. There is a
recommended Vulkan API for querying how many slots will be taken for a
given AHB, but it's unclear how to populate the parameters to be sure of
getting the correct answer. As the spec for that API says that the
maximum number of slots possibly taken would be the maximum number of
planes of a given format that will be sampled via YCbCr sampling, we
simply hardcode that maximum number here.

Bug: dawn:2476
Change-Id: Ie567ca7efc19fe908e2b6e9ab862ee19332820b0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/208274
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>