[glsl] Convert `select` to `mix` instead of ternary.
The current GLSL IR and AST backends translate the `select` builtin to a
ternary condition. This isn't correct as in WGSL both branches need to
be evaluated all of the time while in GLSL only the selected branch of
a ternary is evaluated.

This Cl converts the `select` calls to a `mix` call. In the case of a
WGSL select with vector values and a scalar boolean condition the scalar
boolean is splatted into a vector of correct size.

Bug: 367066331
Change-Id: Ibe2bd5a3aabd29a108831e3ed4469eb40c490278
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/208194
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>