[glsl][ir] Split uniform texture code from texture polyfill.
This CL breaks the texture polyfill into two pieces. The
`TextureBuiltinsFromUniform` and `TexturePolyfill`. This is due to Dawn
sending pre-remapping data for the texture uniform bindings and post
mapping data for the combined samplers transform. In order to avoid
changing Dawn at this state, we're splitting the transform.

Bug: 42251044
Change-Id: I35220d3bd2faff03ab3f3ec3c165084057ffeac5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/209034
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>