HLSL-IR: implement LocalizeStructArrayAssignment for FXC
Implements the same transform as on the AST path. Works around FXC's
compilation failure:

error X3500: array reference cannot be used as an l-value; not natively addressable

This happens in a few cases, but with this transform, we deal with the
case of dynamic indexing an array within a struct local/private
variable.

Bug; 368355095

Change-Id: I2e3c68679afa1cc1ad51fc0e104dca1af9540272
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/208614
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: Antonio Maiorano <amaiorano@google.com>