Remove SwapChain from webgpu.h
This has been completely replaced by WGPUSurface.

Also removes tests, wire support, and dawn::native support for
swapchains. The SwapChain objects stays in dawn::native's implementation
details as folding it in surface would be a lot of changes for this CL.
It is a low-priority cleanup to do so and the current set-up might even
be a good long-term one if we rename SwapChain to SurfaceConnection for
example.

Bug: 42241264
Change-Id: I8e32e9ccceec0a353edd5ac8f0983a9faaff1396
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/208994
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>