[glsl][ir] Add struct padding members
This CL adds the needed explicit padding to structs which are
host-shareable. This is done in the printer to ease implementation, we
inject the new members when emitting the struct, and then we emitting a
constant or constructed struct we inject the padding members. Other
access work as expected as the member indices don't change.

Bug: 42251044
Change-Id: Id81a9acab917b991d1a8b55e9b78b34ee6237d6d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/209874
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>