[dawn][graphite] Add a check for MTLFunction being nil.
There is a crash report which seems that point at the Metal backend
failing to get an MTLFunction after making the MTLLibrary. It's not
clear why that would happen, so add an explicit check that provides more
context which will end up in the CrashKey.

Bug: 372181030
Change-Id: I2404da13dbde2390ffc7b63567eadbdc54dc078e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/210495
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>