[dawn][webgpu.h] Remove webgpu_cpp.h's use of memcpy.
This use was added for a temporary workaround of const char* in
callbacks instead of wgpu::StringView. The addition of the memcpy should
have come with a corresponding #include <cstring> but instead replace
the memcpy with a manually written loop. This is ok because 1) the code
is cold and the deprecated path and 2) compilers know to replace such
simple loops with memcpy calls. This way in the future when the
deprecation workaround is removed, we don't need to remember to remove
the #include.

Bug: 42241188
Change-Id: Ifb6def95790196481797fc9f3962e05ed290d768
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/210934
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Auto-Submit: Corentin Wallez <cwallez@chromium.org>