[hlsl] Fix f16 vector element stores in storage
When the source pointer for an f16 store_vector_element instruction
did not come from a `access` instruction, we were creating an invalid
bitcast from f16 to u32.

Use the `Store()` method on this codepath instead, which takes care of
the special-casing needed for the f16 case.

Bug: 42251045
Change-Id: I3be490f0709000f1acfd3cea844ea3549f7ab6f4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/211095
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Antonio Maiorano <amaiorano@google.com>