[tint] Minor change for fix tests
Fix tests will not be able to find an anchor string if the testing code
(the c++) has a longer match than the (empty) literal string.

As writting this is not a case that can be handled by fix-tests and we
fall back to simply reporting an error.

Change-Id: Ia156be02ac7177c9d76769174b469899eba3f976
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/211535
Commit-Queue: Peter McNeeley <petermcneeley@google.com>
Reviewed-by: James Price <jrprice@google.com>