[ir] Add ReferencedFunctions helper
A new helper that lazily determines (and caches) the set of functions
that are transitively referenced by a function. This is similar to
ReferencedModuleVars.

This will be needed for the SingleEntryPoint transform, which needs to
know which functions are used by a given entry point.

Bug: 374972031
Change-Id: I0fd20bf2e87665dc18cb10725b6aab023eae5ed3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/211795
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>