tint: Make Command::LookPath also look in the executable directory
Added a utility to retrieve the executable path and directory, and make
Command::LookPath also look in there if not found in the CWD, and before
looking in the path environment variable. This should make it a lot
easier to run external executables and dlls as won't have to make sure
to set the path each time.

Change-Id: I09d6ff0a07f1a7904787d4be2a4abdf119a19ad6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/212914
Commit-Queue: Antonio Maiorano <amaiorano@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>