OpenGL: fix placeholder sampler in vertex stage.
The pipeline queries each shader stage in turn to ask if it needs
a placeholder sampler (e.g., for textureLoad() with no sampler).
However, CompileShader() was initializing "needsPlaceholderSampler" to
false on each call. If the vertex stage needs a placeholder sampler,
but the fragment stage doesn't, it will be overwritten back to false.

The fix is not to initialize it to false in CompileShader(). It's
already initialized to false by PipelineGL::InitializeBase().

(Long-winded explanation for a one-line fix.)

Bug: 374890770
Change-Id: I98109e2006c12149be1b6eedaaf05fed7c1c5541
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/213054
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Stephen White <senorblanco@chromium.org>