[tint][ir][fuzz] Disable encoding/decoding SubgroupMatrix in IR binary
This feature isn't fully implemented yet in Tint, so is not ready for
fuzzing. This prevents the fuzzer from attempting to create IR that
touches it. It should be trivial to revert this change in the future
when it is ready for fuzzing.

Fixes: 376278212
Change-Id: I2bc40de99e63f2e1ad007f308e3dfc693aaf13d4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/213134
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>