[dawn][Metal] Disable timestamp sampling on all Intel GPUs
And not just Iris 655 Plus as done previously.

On a variety of Intel GPUs on macOS, if timestamp queries are used at
least once, the GPU gets stuck in high-power mode until the browser is
closed. This causes heating and battery drainage of user's devices so
instead prevent using timestamp queries on these devices.

Bug: 368256332
Change-Id: Ia5c0eb2963fc1378d421c0ab7ca51ef91acee011
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/212896
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>