HLSL-IR: Fix default-only-switch workaround for FXC
Existing implementation on the IR path would add a "case 0" along with
the default case, but did doesn't fix the FXC code gen bug. This patch
implements the same fix we did on the AST path: that is, to replace the
switch with a single-iteration loop.

Also opportunistically renamed FxcPolyfill to ReplaceDefaultOnlySwitch
as this was the only thing it handles, and we've decided to keep each
FXC-specific transform separate.

Bug: 368660972
Bug: 42250197
Change-Id: I9ad11fb089d792c3e6c9ce7be5eccae7fc52f2d2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/213014
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: James Price <jrprice@google.com>