HLSL-IR: fix ShaderIO OOB indexing when mixing subgroup and workgroup inputs with others
GetInput was indexing OOB on input_indices when indexing these other
inputs. For example, for a compute shader with a subgroup parameter
followed  by a regular invocation id, GetInput would intercept the
subgroup parameter to emit the proper Wave call, but for the invocation
id, it would index input_indices at a member index that was out of
bounds. The fix is to make sure input_indices is properly populated,
even for members we won't be looking up in there.

Bug: 363199902
Bug: 357896924
Change-Id: I5e3d30980a10b8beeccf5461611ea96b5bcda7b6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/213379
Auto-Submit: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>