[Vulkan] Allow internal errors when calling GetAHardwareBufferProperties
This method makes a Vulkan API call that will return an error if
`handle` is invalid. This is not cause to lose the Dawn device, as it is
a client-side error and not a true internal Dawn error (see [1] for
fuller analysis).

[1] https://issues.chromium.org/u/1/issues/346282342#comment20

Change-Id: Id0abb3f44e4abc369c59bc8a330c425b8317c4ea
Bug: 346282342
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/213274
Commit-Queue: Colin Blundell <blundell@chromium.org>
Auto-Submit: Colin Blundell <blundell@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>