[glsl] Use GenerateBindings() for texture-builtins-from-uniform
This avoids duplicating the code for generating this across the main
Tint exe, the benchmark, and the fuzzer.

Since GenerateBindings operates on an IR module, we do not have an
Inspector to get the texture queries. Instead, just adding entries for
every texture in the module.

The benchmark can now just generate one set of options for all entry
points, so simplify that logic a bit too.

Bug: 375388105
Change-Id: I9e5685d0bf9ac20c99a08a54faa6cc5cbce92221
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/213377
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>