[glsl] Use SingleEntryPoint in fuzzer
Instead of skipping modules that have multiple entry points, use the
SingleEntryPoint transform to strip the module down to the first entry
point.

Ideally we would generate for each entry point in the module, but we'd
need to be able to clone the whole module to do this, since the
backend mutates it when it runs.

Bug: 375388101
Change-Id: Idf56b5780ac2b54a4ade8a3ba0067f9bb2a6407e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/212534
Reviewed-by: dan sinclair <dsinclair@chromium.org>