[dawn] Support additional 1-component vertex formats
As well as unorm8x4-bgra. These formats weren't added to WebGPU earlier
as they were not available on macOS 10.13 and below. They are now
guaranteed on all platforms that support WebGPU and we can add them.

The Tint vertex pulling transform is updated to support the new formats,
but unorm8x4-bgra will require an additional transform for GLSL that
swizzles some of the attributes.

Bug: 376924407
Change-Id: I9779a5b732528920e12c55d16a3507e78e94481e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/213435
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>