[dawn] Add apparentSize to ExternalTextureDescriptor (noop)
This will be used to support the cases where in JS we have a VideoFrame
that has a visibleRect that doesn't have the same size as the
displaySize. The WGSL textureLoad and textureSize builtins on external
textures currently use the visibleRect size but they should use the
displaySize.

From the Dawn side, displaySize doesn't have a clear semantic, so call
it apparentSize as it's the size that the texture_external appears to be
in WGSL.

Likewise rename visibleOrigin/Size to cropOrigin/Size as it is more
clear what that means in the context of Dawn where we crop plane0.

In future CLs the new ExternalTextureDescriptor attributes will be used
to compute arguments in the ExternalTexture uniform buffer but first we
need Chromium to use them. So this CL only does dawn.json changes.

Bug: 377574981
Change-Id: I5ca5c67b37bf1c373fd9d41638db29cd6ede903c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/213975
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>