[tint][ir][val] Add basic const checks for @workgroup_size
This checks that the value for params are in range if they are
constant values, and rejects non-constant value params if
kAllowsOverrides is not set.

It does not do the look-up to determine if more complex cases are
override-expression yet, since there is still implementation side work
needed for overrides.

Issue: 376624999
Change-Id: I370c982d503e32e6db03f5b0ef91dee5bfe2eb75
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/214095
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>