HLSL-IR: fix DecomposeUniformAccess for vecN<f16>
The old f16 code was not taking into account the size of f16 is 4
instead of 8 bytes. So, the indices and swizzling were all working
on the wrong compoennts. The fix is to handle the f16's as 4 bytes
and use the correct swizzling.

Bug: 359272864
Change-Id: I311a45c187f9de4181d234474fda72a2bc616f23
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/214134
Reviewed-by: David Neto <dneto@google.com>
Commit-Queue: Antonio Maiorano <amaiorano@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>