[dawn] Make EncodingContext::TryEncode fail early on previous errors
All encoder calls go through TryEncode that first checks the status of
the command encoder before calling the code to actually validate/encode
commands. TryEncode did not early-out when an error previously happened.

This appeared to work because any error would cause the encoder.Finish()
to return an error object, so we would just be encoding useless
commands, which is harmless.

However RenderPass.End() performs some validation steps and closing the render
pass involves multiple steps. If validation failures happened in the middle of
them, it would be possible to stay in a partially open state, causing
issues where CommandBufferStateTracker is used while partially working.

This CL adds a new Status to EncodingContext for ErrorInRecording, such
that all future calls to TryEncode early-out. There is some slight
complexity because we need to promote ErrorInRecording to Finished to
preserve the WebGPU semantics of doing device-errors when encoders are
finished.

A regression test is added, and two tests updated that fail because they
expected errors on invalid encoders to become device errors, which is no
longer the case, and more in line with the WebGPU spec.

Bug: 377530684
Change-Id: I3a31c6de636ba7cdc63d54a074671bb19a55c9fb
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/214195
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>