[msl][ast] Emit struct name before initializer
This is needed when a struct initializer is nested inside another
composite (e.g. an array).

This isn't currently an issue for this backend as we hoist all
initializers for other reasons, but the next patch will reintroduce
some nested initializers when constructing packed matrices for
workgroup storage.

Change-Id: I4b77730a88730563bdd7d231fb25390f0f42a057
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/215074
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>