[tint] Improve uniformity diagnostics for builtin parameters
When a builtin function's parameter is required to be uniform, make
sure we actually state that in the error message instead of jumping to
the source of non-uniformity.

This only affects workgroupUniformLoad in this patch, but will also
affect the `subgroupShuffle*` builtins in the next patch.

Change-Id: I56199f5e311ae65ae00e5ee838aefc5c84557064
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/214895
Reviewed-by: David Neto <dneto@google.com>
Commit-Queue: James Price <jrprice@google.com>