[dawn] Support apparentSize in ExternalTexture
Instead of using visibleSize to scale the loadTransform in
ExternalTexture, use apparentSize to allow some scaling to happen.

This CL also changes all uses of visibleSize/Origin to cropSize/Origin
and updates the name in the respective Tint transform.

Validation tests and execution tests are added to check the behavior of
apparentSize.

Bug: 377574981
Change-Id: I7753a18ce190e10b5575b1acdda6ee200f81d7be
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/213994
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>