HLSL: combine two maps into one in PixelLocalOptions
Instead of two unorderd_maps that map a member index to the attachment
index and texel format respectively, we now have a single map that maps
the member index to a struct of both. This avoids having to ensure that
both maps contain the same keys, which is not only simpler, but
increases the chance that fuzzers will produce valid options.

Bug: 380271246
Change-Id: Ic31bb31b19b455ef03996d44ff8f85f57ee91363
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/216154
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Antonio Maiorano <amaiorano@google.com>