Break apart `WriteFile` helper.
The write file helper has a bunch of checks for if we're emitting to
`stdout` or a file. The majority of the code is different between the
two paths. This CL splits it into two `Impl` methods which are called by
`WriteFile` based on if the output is to stdout or not.

Change-Id: I5aa21acd8c6e4ea6fcd0cd93c36e05dbd2122ec5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/216495
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>